+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/common/Tooltip.kt
@@ -103,6 +103,7 @@ fun Tooltip(
                 hideTooltip = hideTooltip,
             )
             .tooltipAnchorSemantics(i18n.commonShowTooltip(), enabled, tooltipState, scope),
+        // TODO: Tooltips should be adaptive (top, bottom, etc. depending on location)
         positionProvider = TooltipDefaults.rememberPlainTooltipPositionProvider(),
         tooltip = {
             TooltipSurface(