+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/room/timeline/element/state/PowerLevelTimelineElementView.kt
@@ -1,0 +1,100 @@
+package de.connect2x.messenger.compose.view.room.timeline.element.state
+
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.ClipEntry
+import de.connect2x.messenger.compose.view.room.timeline.Indicator
+import de.connect2x.messenger.compose.view.room.timeline.IndicatorText
+import de.connect2x.messenger.compose.view.room.timeline.element.TimelineElementView
+import de.connect2x.messenger.compose.view.room.timeline.element.message.bubble.ReferencedMessagePill
+import de.connect2x.trixnity.messenger.viewmodel.room.timeline.elements.BaseTimelineElementHolderViewModel
+import de.connect2x.trixnity.messenger.viewmodel.room.timeline.elements.TimelineElementHolderViewModel
+import de.connect2x.trixnity.messenger.viewmodel.room.timeline.elements.state.PowerLevelsTimelineElementViewModel
+import kotlinx.coroutines.flow.filterNotNull
+import kotlinx.coroutines.flow.first
+import kotlin.reflect.KClass
+
+interface PowerLevelTimelineElementView : TimelineElementView<PowerLevelsTimelineElementViewModel>
+
+
+class PowerLevelTimelineElementViewImpl : PowerLevelTimelineElementView {
+    override val supports: KClass<PowerLevelsTimelineElementViewModel> = PowerLevelsTimelineElementViewModel::class
+
+    override suspend fun waitFor(element: PowerLevelsTimelineElementViewModel) {
+        element.changeMessage.filterNotNull().first()
+    }
+
+    override fun isFocusable(): Boolean = false
+
+    @Composable
+    override fun createInTimeline(
+        holder: BaseTimelineElementHolderViewModel,
+        element: PowerLevelsTimelineElementViewModel,
+        index: Int,
+    ) {
+        StateElement(element)
+    }
+
+    @Composable
+    override fun createAsPreview(
+        holder: TimelineElementHolderViewModel,
+        element: PowerLevelsTimelineElementViewModel,
+        index: Int,
+    ) {
+        StateElement(element)
+    }
+
+    @Composable
+    override fun createReplyInTimeline(
+        holder: TimelineElementHolderViewModel,
+        element: PowerLevelsTimelineElementViewModel,
+        modifier: Modifier,
+        interactionSource: MutableInteractionSource
+    ) {
+        ReferencedMessagePill(
+            holder = holder,
+            modifier = modifier,
+            interactionSource = interactionSource,
+            content = {
+                StateElement(element)
+            }
+        )
+    }
+
+    @Composable
+    override fun createReplyInSendMessage(
+        holder: TimelineElementHolderViewModel,
+        element: PowerLevelsTimelineElementViewModel,
+        modifier: Modifier,
+        interactionSource: MutableInteractionSource
+    ) {
+        ReferencedMessagePill(
+            holder = holder,
+            modifier = modifier,
+            interactionSource = interactionSource,
+            content = {
+                StateElement(element)
+            }
+        )
+    }
+
+    @Composable
+    override fun getClipEntry(
+        holder: BaseTimelineElementHolderViewModel,
+        element: PowerLevelsTimelineElementViewModel
+    ): ClipEntry? = null
+
+}
+
+@Composable
+private fun StateElement(element: PowerLevelsTimelineElementViewModel) {
+    val changeMessage = element.changeMessage.collectAsState().value
+    changeMessage?.let {
+        Indicator(MaterialTheme.colorScheme.tertiary) {
+            IndicatorText(it, MaterialTheme.colorScheme.onTertiary)
+        }
+    }
+}