+++ trixnity-messenger/src/commonTest/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/settings/RoomSettingsViewModelTest.kt
@@ -264,6 +264,7 @@ class RoomSettingsViewModelTest {
             onOpenExportRoom = mock(),
             onOpenUserProfile = mock(),
             onOpenMention = { _, _  -> },
+            onOpenPowerLevel = mock(),
         )
     }
 }